pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "openai>=2.1.0",
    "openai-whisper>=20250625",
    "pyannote-audio>=4.0.0",
    "langchain",
    "langchain_openai",
    "dotenv",
    "fastapi[standard]",
    "python-multipart"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"